rf_Array: Seperate check from actual access
This is to discourage users from assuming that
"RF_ARRAY_CHECKED_INDEX_ACCESS" means that a variable will be returned
for users to play with. This is a false assumption.

Instead a better approach would be to check index accesses on a seperate
line before using the unchecked variant as a proper variable.

This approach will be considered for future application of operations
where it is expected for the "function" to return a variable.